first commit
This doesn't do a lot for now, it just displays some numbers on the
screen. Originally I wanted to do it with LVGL, but I faced multiple issues:
1. I wanted to rotate the screen, but the ili9341 driver seems to be buggy -
   upon orientation change it doesn't update the resolution. Third of the
   screen is static noise...
2. The display driver doesn't implement runtime screen rotation, so I couldn't
   change the orientation from LVGL either.
3. I thought I will flip the text then from LVGL by 90 degrees - which sounds
   simple enough. There are even tutorials how do that with canvas element.
   Unfortunately the board doesn't have enough memory to use canvas, which
   also requires the buffer to be duplicated for the rotation.

So the text display is handled by me, with pixel fonts. The extra memory
it requires is 1400 bytes for the buffer (the framebuffer is updated in 70x10 pixel
chunks, with 2 bytes colors), and about 800 bytes for the monochrome character data
(which could be cut down to under 100 bytes).